requirements.txt
selenium
beautifulsoup4
matplotlib
lxml
requests
langdetect
tqdm
sentence-transformers
scikit-learn
numpy
scipy
hdbscan
umap-learn
fastapi
uvicorn
elasticsearch
dash
dash-bootstrap-components
plotly
pandas
python-dotenv
pytest
pytest-cov
pytest-mock
pytest-asyncio
requests-mock
huggingface_hub[hf_xet]